+++ CHANGELOG.md
@@ -4,6 +4,21 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v0.0.13](https://github.com/chark/blender-skunk/compare/v0.0.12...v0.0.13) - 2024-XX-XX
+
+### Added
+
+- Option to choose on which axis to distribute objects on.
+
+### Changed
+
+- Default child name suffix from `Child` to `LOD0`.
+- Increased min and max sizes when distributing objects.
+
+### Fixed
+
+- Collections not being maintained when creating LODs or empty parents.
+
 ## [v0.0.12](https://github.com/chark/blender-skunk/compare/v0.0.11...v0.0.12) - 2024-12-03
 
 ### Added